DAY 6/Math/requirements.txt
langgraph
langchain-google-genai
python-dotenv
word2number